+++ exonum-java-binding/common/src/test/java/com/exonum/binding/common/blockchain/TransactionResultTest.java
@@ -14,14 +14,14 @@
  * limitations under the License.
  */
 
-package com.exonum.binding.common.transaction;
+package com.exonum.binding.common.blockchain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
-import com.exonum.binding.common.transaction.TransactionResult.Type;
+import com.exonum.binding.common.blockchain.TransactionResult.Type;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
