+++ exonum-java-binding/core/src/main/java/com/exonum/binding/blockchain/serialization/TransactionLocationSerializer.java
@@ -19,7 +19,7 @@ package com.exonum.binding.blockchain.serialization;
 import static com.exonum.binding.common.serialization.StandardSerializers.protobuf;
 
 import com.exonum.binding.common.serialization.Serializer;
-import com.exonum.binding.common.transaction.TransactionLocation;
+import com.exonum.binding.common.blockchain.TransactionLocation;
 
 public enum TransactionLocationSerializer implements Serializer<TransactionLocation> {
   INSTANCE;
