+++ exonum-java-binding/core/src/main/java/com/exonum/binding/blockchain/serialization/TransactionResultSerializer.java
@@ -17,12 +17,12 @@
 package com.exonum.binding.blockchain.serialization;
 
 import static com.exonum.binding.common.serialization.StandardSerializers.protobuf;
-import static com.exonum.binding.common.transaction.TransactionResult.MAX_USER_DEFINED_ERROR_CODE;
-import static com.exonum.binding.common.transaction.TransactionResult.SUCCESSFUL_RESULT_STATUS_CODE;
-import static com.exonum.binding.common.transaction.TransactionResult.UNEXPECTED_ERROR_STATUS_CODE;
+import static com.exonum.binding.common.blockchain.TransactionResult.MAX_USER_DEFINED_ERROR_CODE;
+import static com.exonum.binding.common.blockchain.TransactionResult.SUCCESSFUL_RESULT_STATUS_CODE;
+import static com.exonum.binding.common.blockchain.TransactionResult.UNEXPECTED_ERROR_STATUS_CODE;
 
 import com.exonum.binding.common.serialization.Serializer;
-import com.exonum.binding.common.transaction.TransactionResult;
+import com.exonum.binding.common.blockchain.TransactionResult;
 
 public enum TransactionResultSerializer implements Serializer<TransactionResult> {
   INSTANCE;
