+++ exonum-java-binding/core/src/main/java/com/exonum/binding/transaction/TransactionExecutionException.java
@@ -32,7 +32,7 @@ import javax.annotation.Nullable;
  * <a href="https://exonum.com/doc/version/0.10/advanced/node-management/#transaction">the API endpoint documentation</a>
  * for more information.
  *
- * @see com.exonum.binding.common.transaction.TransactionResult
+ * @see com.exonum.binding.common.blockchain.TransactionResult
  */
 public class TransactionExecutionException extends Exception {
