+++ exonum-java-binding/qa-service/src/main/java/com/exonum/binding/qaservice/QaService.java
@@ -20,8 +20,8 @@ import com.exonum.binding.blockchain.Block;
 import com.exonum.binding.common.configuration.StoredConfiguration;
 import com.exonum.binding.common.hash.HashCode;
 import com.exonum.binding.common.message.TransactionMessage;
-import com.exonum.binding.common.transaction.TransactionLocation;
-import com.exonum.binding.common.transaction.TransactionResult;
+import com.exonum.binding.common.blockchain.TransactionLocation;
+import com.exonum.binding.common.blockchain.TransactionResult;
 import com.exonum.binding.service.Service;
 import java.util.List;
 import java.util.Map;
