+++ exonum-java-binding/qa-service/src/test/java/com/exonum/binding/qaservice/ApiControllerIntegrationTest.java
@@ -71,8 +71,8 @@ import com.exonum.binding.common.message.TransactionMessage;
 import com.exonum.binding.common.serialization.json.JsonSerializer;
 import com.exonum.binding.common.serialization.json.TransactionLocationAdapterFactory;
 import com.exonum.binding.common.serialization.json.TransactionResultAdapterFactory;
-import com.exonum.binding.common.transaction.TransactionLocation;
-import com.exonum.binding.common.transaction.TransactionResult;
+import com.exonum.binding.common.blockchain.TransactionLocation;
+import com.exonum.binding.common.blockchain.TransactionResult;
 import com.exonum.binding.service.InternalServerError;
 import com.exonum.binding.test.Bytes;
 import com.google.common.collect.ImmutableMap;
