+++ exonum-light-client/src/main/java/com/exonum/client/response/TransactionResponse.java
@@ -17,8 +17,8 @@
 package com.exonum.client.response;
 
 import com.exonum.binding.common.message.TransactionMessage;
-import com.exonum.binding.common.transaction.TransactionLocation;
-import com.exonum.binding.common.transaction.TransactionResult;
+import com.exonum.binding.common.blockchain.TransactionLocation;
+import com.exonum.binding.common.blockchain.TransactionResult;
 import lombok.NonNull;
 import lombok.Value;
